Topバーを切り出した
